+++ serapeum-core/src/serapeum/core/configs/configs.py
@@ -3,7 +3,7 @@
 from dataclasses import dataclass
 from typing import Optional
 
-from serapeum.core.llm import LLM
+from serapeum.core.llms import LLM
 from serapeum.core.models import StructuredLLMMode
