+++ serapeum-core/src/serapeum/core/structured_tools/text_completion_llm.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, Optional, Tuple, Type, Union
 from pydantic import BaseModel
 
 from serapeum.core.configs.configs import Configs
-from serapeum.core.llm.base import LLM
+from serapeum.core.llms.base import LLM
 from serapeum.core.output_parsers import BaseParser, PydanticParser
 from serapeum.core.prompts.base import BasePromptTemplate, PromptTemplate
 from serapeum.core.structured_tools.models import BasePydanticLLM
