+++ serapeum-core/src/serapeum/core/structured_tools/tools_llm.py
@@ -17,8 +17,8 @@ from typing import (
 from pydantic import BaseModel
 
 from serapeum.core.configs.configs import Configs
-from serapeum.core.llm.base import LLM
-from serapeum.core.llm.function_calling import FunctionCallingLLM
+from serapeum.core.llms.base import LLM
+from serapeum.core.llms.function_calling import FunctionCallingLLM
 from serapeum.core.prompts.base import BasePromptTemplate, PromptTemplate
 from serapeum.core.structured_tools.models import BasePydanticLLM, Model
 from serapeum.core.structured_tools.utils import StreamingObjectProcessor
