+++ serapeum-core/src/serapeum/core/structured_tools/utils.py
@@ -12,8 +12,8 @@ from serapeum.core.output_parsers import PydanticParser
 
 if TYPE_CHECKING:
     from serapeum.core.base.llms.models import ChatResponse
-    from serapeum.core.llm.base import LLM
-    from serapeum.core.llm.function_calling import FunctionCallingLLM
+    from serapeum.core.llms.base import LLM
+    from serapeum.core.llms.function_calling import FunctionCallingLLM
     from serapeum.core.prompts.base import BasePromptTemplate
     from serapeum.core.structured_tools.models import BasePydanticLLM
     from serapeum.core.tools.models import ToolCallArguments
